Change NPC name hiding behavior
With the setting turned on, hide all NPC names that are not set to 
ALWAYS in their prototype tokens.

This is so you can have friendly NPCs fighting alongside ou with their 
names showing.

Perhaps a better option would be to show messages from tokens not set to 
FRIENDLY only to the GM so you can hide their healings and resistances 
better.